chore: useAuth 훅 개선
